chore: update expandable component documents, add screenshot tests, add template with sample
